requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6
pillow==10.0.0
torch==2.0.1
transformers==4.33.2
numpy==1.25.2
requests==2.31.0
huggingface-hub